Refactor component exports and queries
- Switch to named exports for all components to facilitate better tree-shaking and code maintainability.
- Renamed GraphQL queries to prevent potential conflicts and improve clarity.
- Updated imports across all pages and components to reflect these export changes.
- Cleaned up unused code and comments for a more streamlined codebase.
- Ensured pages importing components use the updated exports.
- Changes are in preparation for further modularization of the code and may require updates to any external references to the modified components.
