+++ src/components/ExternalLink.tsx
@@ -15,9 +15,9 @@ const ExternalLink = ({
   );
 };
 
-export default ExternalLink;
+export { ExternalLink };
 
-export const query = graphql`
+export const externalLinkQuery = graphql`
   fragment ExternalLink on ContentfulExternalLink {
     __typename
     contentful_id
