+++ src/components/index.ts
@@ -1,21 +1,23 @@
-import ExternalLink from './ExternalLink';
-import Header from './Header';
-import Layout from './Layout';
-import ListTags from './ListTags';
-import Pen from './CodePen';
-import Posts from './Posts';
-import Seo from './Seo';
-import SiteMap from './Sitemap';
-import Tags from './Tags';
+import { ExternalLink } from './ExternalLink';
+import { Header } from './Header';
+import { ImageHeader } from './ImageHeader';
+import { Layout } from './Layout';
+import { ListTags } from './ListTags';
+import { Pen } from './CodePen';
+import { Posts } from './Posts';
+import { Seo } from './Seo';
+import { Sitemap } from './Sitemap';
+import { Tags } from './Tags';
 
 export {
   ExternalLink,
   Header,
+  ImageHeader,
   Layout,
   ListTags,
   Pen,
   Posts,
   Seo,
-  SiteMap,
+  Sitemap,
   Tags,
 };
