+++ src/pages/entitlements.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react';
 
-import Layout from '../components/Layout';
-import Seo from '../components/Seo';
-import SiteMap from '../components/Sitemap';
+import { Layout, Seo } from '../components';
 
 const Entitlements = () => (
   <Layout>
